requirements.txt
Jinja2==2.8
PyYAML==3.12
pytest==3.0.4
pexpect==4.2.1
